Use VI algorithm to fit PSD model

Fit PSD model with VI algorithm, and use the loss function as a stopping criterion.

Usage

psd_fit_vi(G, K, epsilon = 1e-05, maxiter = 500)

Arguments

G

The I x J matrix of counts; all entries of G should be taken from {0,1,2}.

K

An integer 2 or greater giving the matrix rank.

epsilon

Convergence criterion.

maxiter

The maximum number of iterations.

Value

A List with the following parameters:

P

The population scale matrix of the individuals.

F

The gene scale matrix of the populations.

Loss

A vector represents the value of the loss function which records once for 10 iterations.

Iterations

An integer represents the number of iterations.

Examples

G <- matrix(c(0,0,1, 0,2,1, 1,0,1, 0,1,0, 1,0,0), 3, 5)
psd_fit_vi(G, 2, 1e-5, 10)
#> $P
#>           [,1]      [,2]
#> [1,] 0.6125485 0.3874515
#> [2,] 0.4543082 0.5456918
#> [3,] 0.5282933 0.4717067
#> 
#> $F
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.2885826 0.4402571 0.4208492 0.2689115 0.3150110
#> [2,] 0.3124824 0.5647165 0.3774814 0.3343761 0.2838648
#> 
#> $Loss
#> [1] -1.789631
#> 
#> $Iterations
#> [1] 10
#>